+++ procedural_city_generation/roadmap/getRule.py
@@ -24,14 +24,14 @@ def getRule(vertex):
     x = (vertex.coords[0]+singleton.border[0])/(singleton.border[0]*2)
     y = (vertex.coords[1]+singleton.border[1])/(singleton.border[1]*2)
 
-    population_density = np.sqrt((singleton.img2[singleton.img2.shape[0]-y*singleton.img2.shape[0]][x*singleton.img2.shape[1]][0])/255)
+    population_density = np.sqrt((singleton.img2[int(singleton.img2.shape[0]-y*singleton.img2.shape[0])][int(x*singleton.img2.shape[1])][0])/255)
     if vertex.seed:
         return (4, None, population_density)
 
 
     if not vertex.minor_road:
         #Finds the relative position of the vertex on the growth_rule_image
-        intrule=np.argmax(singleton.img[singleton.img.shape[0]-y*singleton.img.shape[0]][x*singleton.img.shape[1]])
+        intrule=np.argmax(singleton.img[int(singleton.img.shape[0]-y*singleton.img.shape[0])][int(x*singleton.img.shape[1])])
         z=(0, 0)
 
         #If the rule is radial, find the closest radial center
